+++ src/pages/WkChat.tsx
@@ -1,0 +1,11 @@
+import Layout from '../components/Layout';
+
+const WkChat = () => {
+  return (
+    <Layout>
+      <h1>WkChat</h1>
+    </Layout>
+  );
+};
+
+export default WkChat;
