+++ src/styles/Home.css
@@ -30,3 +30,8 @@
   color: royalblue;
   cursor: pointer;
 }
+
+.individual-page-kengo {
+  color: teal;
+  cursor: pointer;
+}
