+++ src/styles/Layout.css
@@ -1,4 +1,7 @@
 .layout-wrapper {
   width: 100vw;
   height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
